+++ test/Benchmarks/Exporter/JaegerExporterBenchmarks.cs
@@ -15,12 +15,11 @@
 // </copyright>
 
 using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using BenchmarkDotNet.Attributes;
+using OpenTelemetry.Benchmarks;
 using OpenTelemetry.Exporter.Jaeger;
 using OpenTelemetry.Internal;
 using Thrift.Transport;
@@ -42,7 +41,7 @@ namespace OpenTelemetry.Exporter.Benchmarks
         [GlobalSetup]
         public void GlobalSetup()
         {
-            this.activity = CreateTestActivity();
+            this.activity = ActivityHelper.CreateTestActivity();
             this.activityBatch = new CircularBuffer<Activity>(this.NumberOfSpans);
         }
 
@@ -67,77 +66,6 @@ namespace OpenTelemetry.Exporter.Benchmarks
             }
 
             exporter.Shutdown();
-        }
-
-        internal static Activity CreateTestActivity()
-        {
-            var startTimestamp = new DateTimeOffset(2019, 1, 1, 0, 0, 0, TimeSpan.Zero);
-            var endTimestamp = startTimestamp.AddSeconds(60);
-            var eventTimestamp = new DateTimeOffset(2019, 1, 1, 0, 0, 0, TimeSpan.Zero);
-
-            var traceId = ActivityTraceId.CreateFromString("e8ea7e9ac72de94e91fabc613f9686b2".AsSpan());
-            var spanId = ActivitySpanId.CreateFromString("6a69db47429ea340".AsSpan());
-            var parentSpanId = ActivitySpanId.CreateFromBytes(new byte[] { 12, 23, 34, 45, 56, 67, 78, 89 });
-            var attributes = new Dictionary<string, object>
-            {
-                { "stringKey", "value" },
-                { "longKey", 1L },
-                { "longKey2", 1 },
-                { "doubleKey", 1D },
-                { "doubleKey2", 1F },
-                { "boolKey", true },
-            };
-            var events = new List<ActivityEvent>
-            {
-                new ActivityEvent(
-                    "Event1",
-                    eventTimestamp,
-                    new ActivityTagsCollection(new Dictionary<string, object>
-                    {
-                        { "key", "value" },
-                    })),
-                new ActivityEvent(
-                    "Event2",
-                    eventTimestamp,
-                    new ActivityTagsCollection(new Dictionary<string, object>
-                    {
-                        { "key", "value" },
-                    })),
-            };
-
-            var linkedSpanId = ActivitySpanId.CreateFromString("888915b6286b9c41".AsSpan());
-
-            using var activitySource = new ActivitySource(nameof(CreateTestActivity));
-
-            var tags = attributes.Select(kvp => new KeyValuePair<string, object>(kvp.Key, kvp.Value.ToString()));
-
-            var links = new[]
-            {
-                new ActivityLink(new ActivityContext(
-                    traceId,
-                    linkedSpanId,
-                    ActivityTraceFlags.Recorded)),
-            };
-
-            using var listener = new ActivityListener()
-            {
-                ShouldListenTo = a => a.Name == nameof(CreateTestActivity),
-                Sample = (ref ActivityCreationOptions<ActivityContext> a) => ActivitySamplingResult.AllDataAndRecorded,
-            };
-
-            ActivitySource.AddActivityListener(listener);
-
-            Activity activity = activitySource.StartActivity(
-                "Name",
-                ActivityKind.Client,
-                parentContext: new ActivityContext(traceId, parentSpanId, ActivityTraceFlags.Recorded),
-                tags,
-                links,
-                startTime: startTimestamp);
-
-            activity.Stop();
-
-            return activity;
         }
 
         private class BlackHoleTransport : TTransport
